move bareos repo & pkg installs to build, from entrypoint
In all the author matching images we run as bareos:bareos,
which does not have permission to install within the container.
File does run as root:bareos: but a working set of daemons is
the initial goal here. With as much similarity in their build/run
methods as possible.

Includes:
- additional env var re myself.conf Name element config.
- host user expectations/config guidelines.
- invocation examples.
- README.md entry regarding package elements ignored by this image.